+++ Natom.Petshop.Gestion.Backend/Natom.Petshop.Gestion.Entities/DTO/Pedidos/PedidoListDTO.cs
@@ -51,7 +51,7 @@ namespace Natom.Petshop.Gestion.Entities.DTO.Pedidos
         public bool Prepared { get; set; }
 
         [JsonProperty("peso_total_gramos")]
-        public int PesoTotalGramos { get; set; }
+        public decimal PesoTotalGramos { get; set; }
 
         [JsonProperty("fechaHoraPreparado")]
         public DateTime? FechaHoraPreparado { get; set; }
