+++ Natom.Petshop.Gestion.Backend/Natom.Petshop.Gestion.Entities/Model/OrdenDePedidoDetalle.cs
@@ -15,8 +15,8 @@ namespace Natom.Petshop.Gestion.Entities.Model
         public int ProductoId { get; set; }
         public Producto Producto { get; set; }
 
-        public int Cantidad { get; set; }
-        public int? CantidadEntregada { get; set; }
+        public decimal Cantidad { get; set; }
+        public decimal? CantidadEntregada { get; set; }
 
         public int DepositoId { get; set; }
         public Deposito Deposito { get; set; }
