+++ Natom.Petshop.Gestion.Backend/Natom.Petshop.Gestion.Entities/Model/Results/spReportOrdenDePedidoResult.cs
@@ -37,7 +37,7 @@ namespace Natom.Petshop.Gestion.Entities.Model.Results
 		public decimal PesoTotalEnKilogramos { get; set; }
 		public string DetalleCodigo { get; set; }
 		public string DetalleDescripcion { get; set; }
-		public int DetalleCantidad { get; set; }
+		public decimal DetalleCantidad { get; set; }
 		public decimal DetallePesoUnitarioEnKilogramos { get; set; }
 		public decimal? DetallePrecioUnitario { get; set; }
 		public decimal DetallePesoTotalEnKilogramos { get; set; }
