+++ Natom.Petshop.Gestion.Backend/Natom.Petshop.Gestion.Entities/Model/Results/spStockListaReportResult.cs
@@ -13,8 +13,8 @@ namespace Natom.Petshop.Gestion.Entities.Model.Results
         public string Deposito { get; set; }
         public string Categoria { get; set; }
         public string Producto { get; set; }
-        public int Confirmado { get; set; }
-        public int Reservado { get; set; }
-        public int Real { get; set; }
+        public decimal Confirmado { get; set; }
+        public decimal Reservado { get; set; }
+        public decimal Real { get; set; }
     }
 }
