+++ Natom.Petshop.Gestion.Backend/Natom.Petshop.Gestion.Entities/Model/Results/spVentasRepartoVsMostradorReportResult.cs
@@ -12,13 +12,13 @@ namespace Natom.Petshop.Gestion.Entities.Model.Results
     {
         public string Codigo { get; set; }
         public string Descripcion { get; set; }
-        public int CantidadMostrador { get; set; }
+        public decimal CantidadMostrador { get; set; }
         public decimal VariacionCantidadMostrador { get; set; }
         public string ColorVariacionCantidadMostrador { get; set; }
         public decimal KilosMostrador { get; set; }
         public decimal VariacionKilosMostrador { get; set; }
         public string ColorVariacionKilosMostrador { get; set; }
-        public int CantidadReparto { get; set; }
+        public decimal CantidadReparto { get; set; }
         public decimal VariacionCantidadReparto { get; set; }
         public string ColorVariacionCantidadReparto { get; set; }
         public decimal KilosReparto { get; set; }
