+++ Natom.Petshop.Gestion.Backend/Natom.Petshop.Gestion.Entities/Model/Venta.cs
@@ -31,7 +31,7 @@ namespace Natom.Petshop.Gestion.Entities.Model
 
 		public decimal MontoTotal { get; set; }
 
-		public int PesoTotalEnGramos { get; set; }
+		public decimal PesoTotalEnGramos { get; set; }
 
 		public List<VentaDetalle> Detalle { get; set; }
